Set default values for dates when creating new joblisting
