+++ app/routes/events/EventCreateRoute.js
@@ -7,7 +7,8 @@ import { createEvent } from 'app/actions/EventActions';
 import { uploadFile } from 'app/actions/FileActions';
 import EventEditor from './components/EventEditor';
 import { LoginPage } from 'app/components/LoginForm';
-import { transformEvent, time } from './utils';
+import { transformEvent } from './utils';
+import { time } from 'app/utils/time.js';
 import replaceUnlessLoggedIn from 'app/utils/replaceUnlessLoggedIn';
 
 const mapStateToProps = (state, props) => {
