+++ app/routes/joblistings/JoblistingCreateRoute.js
@@ -7,11 +7,15 @@ import JoblistingEditor from 'app/routes/joblistings/components/JoblistingEditor
 import { push } from 'react-router-redux';
 import { LoginPage } from 'app/components/LoginForm';
 import replaceUnlessLoggedIn from 'app/utils/replaceUnlessLoggedIn';
+import { time } from 'app/utils/time.js';
 
 const mapStateToProps = () => ({
   initialValues: {
     text: '',
     description: '',
+    visibleFrom: time({ hours: 12 }),
+    visibleTo: time({ days: 31, hours: 23, minutes: 59 }),
+    deadline: time({ days: 30, hours: 23, minutes: 59 }),
     fromYear: 1,
     toYear: 5,
     jobType: 'summer_job',
